+++ packages/terra-brand-footer/src/terra-dev-site/test/brand-footer/VerticalMixedHeadersBrandFooter.test.jsx
@@ -101,4 +101,5 @@ export default () => (
         <small>Copyright &copy; 2018 Cerner Corporation.  All Rights Reserved.</small>
       </span>
 )}
-  />);
+  />
+);