+++ packages/terra-infinite-list/src/terra-dev-site/doc/example/InfiniteListExample.jsx
@@ -4,9 +4,10 @@ import LoadingOverlay from 'terra-overlay/lib/LoadingOverlay';
 import OverlayContainer from 'terra-overlay/lib/OverlayContainer';
 import Arrange from 'terra-arrange';
 import classNames from 'classnames/bind';
+import debounce from 'lodash.debounce';
 /* eslint-disable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 import styles from 'terra-infinite-list/lib/terra-dev-site/doc/example/InfiniteListExample.scss';
-import InfiniteList from 'terra-infinite-list/lib/InfiniteList';
+import InfiniteList, { Item } from 'terra-infinite-list/lib/index';
 /* eslint-enable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 
 const cx = classNames.bind(styles);
@@ -15,14 +16,15 @@ class InfiniteListExample extends React.Component {
   constructor(props) {
     super(props);
 
-    this.addMoreData = this.debounce(this.addMoreData.bind(this), 1000);
-    this.handleOnChange = this.handleOnChange.bind(this);
+    this.addMoreData = debounce(this.addMoreData.bind(this), 1000);
+    this.handleSelection = this.handleSelection.bind(this);
 
-    this.state = { stillLoading: true, numberOfPages: 0, selectedIndexes: [] };
+    this.state = { stillLoading: true, numberOfPages: 0, selectedKey: null };
   }
 
-  handleOnChange(event, index) {
-    this.setState({ selectedIndexes: [index] });
+  handleSelection(event, metaData) {
+    event.preventDefault();
+    this.setState({ selectedKey: metaData.key });
   }
 
   addMoreData() {
@@ -35,23 +37,17 @@ class InfiniteListExample extends React.Component {
     this.setState(newState);
   }
 
-  debounce(fn, delay) {
-    let timer = null;
-    return (...args) => {
-      const context = this;
-      clearTimeout(timer);
-      timer = setTimeout(() => {
-        fn.apply(context, args);
-      }, delay);
-    };
-  }
-
   render() {
     const items = [];
     for (let i = 0; i < 15 * this.state.numberOfPages; i += 1) {
-      items.push(<InfiniteList.Item
-        key={`${i}`}
-        content={(
+      items.push(
+        <Item
+          key={`${i}`}
+          isSelected={this.state.selectedKey === `item-${i}`}
+          isSelectable
+          onSelect={this.handleSelection}
+          metaData={{ key: `item-${i}` }}
+        >
           <Arrange
             fitStart={<h3 style={{ width: '50px' }}>{`${i}`}</h3>}
             fill={(
@@ -62,13 +58,13 @@ class InfiniteListExample extends React.Component {
                   Donec tempor mi vitae lorem congue, ut ultrices metus feugiat. Sed non commodo felis.
                   Aliquam eget maximus dui, ut rhoncus augue.
               </p>
-)}
+            )}
             fitEnd={<div className={cx(['icon-wrapper'])}><IconInformation /></div>}
             align="center"
             fitStartAttributes={{ style: { textAlign: 'center' } }}
           />
-)}
-      />);
+        </Item>,
+      );
     }
 
     const fullLoading = <LoadingOverlay isOpen isAnimated isRelativeToContainer backgroundStyle="dark" />;
@@ -85,10 +81,7 @@ class InfiniteListExample extends React.Component {
       }}
       >
         <InfiniteList
-          isSelectable
-          isDivided
-          onChange={this.handleOnChange}
-          selectedIndexes={this.state.selectedIndexes}
+          dividerStyle="standard"
           isFinishedLoading={!this.state.stillLoading}
           onRequestItems={this.addMoreData}
           initialLoadingIndicator={fullLoading}