+++ packages/terra-popup/src/terra-dev-site/doc/example/PopupBounded.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Button from 'terra-button';
 /* eslint-disable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 import Popup from 'terra-popup/lib/Popup';
-import ExamplePopupContent from 'terra-popup/lib/terra-dev-site/doc/common/ExamplePopupContent';
+import Placeholder from 'terra-doc-template/lib/Placeholder';
 /* eslint-enable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 
 class PopupBounded extends React.Component {
@@ -10,9 +10,10 @@ class PopupBounded extends React.Component {
     super(props);
     this.handleButtonClick = this.handleButtonClick.bind(this);
     this.handleRequestClose = this.handleRequestClose.bind(this);
-    this.handleOnChange = this.handleOnChange.bind(this);
     this.setParentNode = this.setParentNode.bind(this);
     this.getParentNode = this.getParentNode.bind(this);
+    this.setButtonNode = this.setButtonNode.bind(this);
+    this.getButtonNode = this.getButtonNode.bind(this);
     this.state = { open: false };
   }
 
@@ -24,15 +25,19 @@ class PopupBounded extends React.Component {
     return this.parentNode;
   }
 
+  setButtonNode(node) {
+    this.parentNode = node;
+  }
+
+  getButtonNode() {
+    return this.parentNode;
+  }
+
   handleButtonClick() {
     this.setState({ open: true });
   }
 
   handleRequestClose() {
-    this.setState({ open: false });
-  }
-
-  handleOnChange() {
     this.setState({ open: false });
   }
 
@@ -52,12 +57,12 @@ class PopupBounded extends React.Component {
           contentWidth="320"
           isOpen={this.state.open}
           onRequestClose={this.handleRequestClose}
-          targetRef={() => document.getElementById('popup-bounded')}
+          targetRef={this.getButtonNode}
           isContentFocusDisabled
         >
-          <ExamplePopupContent onChange={this.handleOnChange} />
+          <Placeholder title="Popup Content" />
         </Popup>
-        <Button id="popup-bounded" text="Bounded Popup" onClick={this.handleButtonClick} />
+        <Button text="Bounded Popup" onClick={this.handleButtonClick} refCallback={this.setButtonNode} />
       </div>
     );
   }