+++ packages/terra-application-links/CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased
 ### Changed
 * Update external links example to point to terra-ui.com
 
+### Fixed
+* Fixed TabMenuList styling.
+
 5.2.0 - (February 13, 2019)
 ------------------
 ### Changed